requirements-webui.txt
# Web UI Backend Requirements
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-socketio>=5.10.0
python-multipart>=0.0.6
pydantic>=2.5.0
sqlalchemy>=2.0.23
alembic>=1.13.0
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0

# Existing project dependencies (if not already installed)
pandas>=2.1.0
numpy>=1.24.0
requests>=2.31.0
aiofiles>=23.2.1
psutil>=5.9.0